feat(basic): add first-lesson.php, get-users-controller.php, and index.view.php files
Add first-lesson.php file with a for loop, a foreach loop, a while loop, and examples of array_filter and array_map functions.

Add get-users-controller.php file with an array of users and require index.view.php file.

Add index.view.php file with a header and a list of users.